refactor(policies): extract PolicyEditorToolbar and color picker
Move TipTap toolbar config, selection sync, and color popover into co-located components to shrink PolicyEditorPage.
